requirements.txt
google-genai
python-dotenv
Pillow
pygame